requirements.txt
reportlab>=4.0.0
pytesseract==0.3.10
pdf2image==1.16.3
pdfplumber==0.9.0
spacy==3.5.3
pandas==2.0.3
python-dotenv==1.0.0
pytest>=8.3.5
pytest-cov>=6.0.0
pypdf>=3.0.0  # Instead of PyPDF2
python-dateutil>=2.8.2
python-magic>=0.4.27
pdfplumber>=0.7.0
numpy>=1.21.0
scikit-learn>=1.0.0
spacy>=3.0.0
